try not do collective calls for vtkhdf
