+++ src/diagnostic/detail/vtk_types/fluid.hpp
@@ -283,7 +283,7 @@ void FluidDiagnosticWriter<H5Writer>::write(DiagnosticProperties& diagnostic)
         [&](auto const& level) {
             auto const ilvl = level.getLevelNumber();
 
-            VTKFileWriter writer{diagnostic, this, info.offset_per_level[ilvl]};
+            VTKFileWriter writer{diagnostic, this, info.offset_per_level[ilvl], ilvl};
 
             auto const write_quantity = [&](auto& layout, auto const&, auto const) {
                 if constexpr (solver::is_hybrid_model_v<Model_t>)
